made the values private, added getters and setters, put the test in a separate classÐ
